C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik4.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik4.cpp
@@ -100,9 +100,9 @@ void MoveBall()
       	int s = Dx/d; // sin
       	int e = Dy/d; // cos
       	if (d < r1+r2){
-      		int Vn1 = vx2*s + vy2*e;
-      		int Vn2 = vx1*s + vy1*e;
-      		int dt = (r2+r1-d)/(Vn1-Vn2);
+      		float Vn1 = vx2*s + vy2*e;
+      		float Vn2 = vx1*s + vy1*e;
+      		float dt = (r2+r1-d)/(Vn1-Vn2);
       		if (dt > 0.6) dt = 0.6;
       		if (dt < -0.6) dt = -0.6;
       		x1-=vx1*dt;
